FIXED: Moved the class type part of the grammar above the class section to keep class types from being hijacked.
Added ':' as a possible end marker for methods so things are handled properly in class
type definitions. Note, however that this causes the method name in the type definition
to also appear in the symbol list, but I'm not sure that's bad.

Corrected the spelling of tuple.

Changed the end pattern for meta.object.ocaml to allow for optional ;; after the 'end'.

Changed tag on the input and really_input snippets to inp and rinp respectively to
avoid a collision with the trigger for stdin.

TODO (anyone , feel free to take these as I may be a while getting around to them):

Need to fix the tuple definition, as it is catching constructs like:
let f (a : int) b = ...
and labeling the (a : int) as a tuple.  Also, the int is being considered a parameter
of the function and not a type, as it should be.  This might be fixed by matching on,
say (\()([^,],)+(\)), but match patterns can't have sub-patterns, so something
slightly different is needed.

Folding markers are messed up for classes and modules... this needs fixing if possible.

Class parameters (e.g. the xx, yy, and zz in class blah xx yy zz = ...) aren't given
any scope other than source.ocaml.  This may or may not be worth fixing.

QUESTIONS:
Folding markers for let function bindings are also pretty silly.  I'm wondering if
these shouldn't just be removed, as they are next to useless, and do little more than
add extraneous fold start markers to the window.  Does anyone else agree?

Why are true and false labeled as "constant.language.pseudo-variable.ocaml"?
Wouldn't constant.boolean.ocaml be better?

William D. Neumann



git-svn-id: http://svn.textmate.org/trunk/Bundles/OCaml.tmbundle@4777 dfb7d73b-c2ec-0310-8fea-fb051d288c6d